requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pyotp==2.9.0
pycryptodome==3.21.0
n